+++ src/schemathesis/specs/openapi/adapter/__init__.py
@@ -1,4 +1,4 @@
-from schemathesis.specs.openapi.adapter import v2, v3
+from schemathesis.specs.openapi.adapter import v2, v3_0, v3_1
 from schemathesis.specs.openapi.adapter.parameters import prepare_parameters
 from schemathesis.specs.openapi.adapter.responses import OpenApiResponse, OpenApiResponses
 
@@ -7,5 +7,6 @@ __all__ = [
     "OpenApiResponses",
     "prepare_parameters",
     "v2",
-    "v3",
+    "v3_0",
+    "v3_1",
 ]